Fix For no Release Event Received
When certain custom actions like showing a QMenu
are done in a slot connected to the groupRightClicked()
signal which gets emitted in qanGroupItem::mousePressEvent()
the release event of the mouse button is never handled
by the QTQuick framework and e.g. the grabber item is
not cleared and will be re-used in the next mouse press
event.

This is fixed by moving the emit of 'groupRightClicked()'
to the mouseReleaseEvent() handler.